#3 Refactored data getters
