--- frontend/src/components/hooks/use-get-politicians.tsx
@@ -1,31 +1,0 @@
-import { useEffect, useState } from "react";
-
-export interface Politician {
-  id: number;
-  firstName: string;
-  lastName: string;
-}
-
-export const useGetPoliticians = (): [Politician[], () => void] => {
-  const [politicians, setPoliticians] = useState<Politician[]>([]);
-  const fetchQuery = () => {
-    (async () => {
-      const searchResult = await fetch(
-        `${
-          process.env.PROXY_ENDPOINT || "http://167.99.244.228:5300"
-        }/politicians`,
-        {
-          mode: "cors",
-        }
-      ).then((response) => response.json());
-      const politiciansResult = searchResult.data.politicians;
-      if (politiciansResult) {
-        setPoliticians(politiciansResult);
-      }
-    })();
-  };
-  useEffect(() => fetchQuery(), []);
-  return [politicians, fetchQuery];
-};
-
-export default useGetPoliticians;
